Move node definitions into seperate module
